+++ src/main/java/io/coodoo/workhorse/jobengine/boundary/JobContext.java
@@ -57,7 +57,7 @@ public class JobContext {
 
     /**
      * Adds a timestamp followed by the message text in as a new line to the executions log <br>
-     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code><br>
+     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
      * Example: <code>[22:06:42.680] Step 3 complete</code>
      * 
      * @param message text to log
@@ -67,8 +67,10 @@ public class JobContext {
     }
 
     /**
-     * Adds a timestamp followed by the message text in as a new line to the executions log and also adds the message in severity INFO to the server log<br>
-     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code><br>
+     * Adds a timestamp followed by an info marker and the info message text in as a new line to the executions log and also adds the message in severity INFO
+     * to the server log<br>
+     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
+     * Info marker: Only if defined in {@link JobEngineConfig#LOG_INFO_MARKER}<br>
      * Example: <code>[22:06:42.680] Step 3 complete</code>
      * 
      * @param logger server log logger
@@ -76,14 +78,26 @@ public class JobContext {
      */
     public void logInfo(Logger logger, String message) {
         logger.info(message);
+        logInfo(message);
+    }
+
+    /**
+     * Adds a timestamp followed by an info marker and the info message text in as a new line to the executions log<br>
+     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
+     * Info marker: Only if defined in {@link JobEngineConfig#LOG_INFO_MARKER}<br>
+     * Example: <code>[22:06:42.680] Step 3 complete</code>
+     * 
+     * @param message text to log
+     */
+    public void logInfo(String message) {
         appendLog(message, true, "i");
     }
 
     /**
      * Adds a timestamp followed by an warn marker and the warn message as a new line to the executions log. It also adds the message in severity WARN to the
      * server log<br>
-     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code><br>
-     * Error marker: <code>[WARN]</code><br>
+     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
+     * Error marker: <code>[WARN]</code> or as defined in {@link JobEngineConfig#LOG_WARN_MARKER}<br>
      * Example: <code>[22:06:42.680] [WARN] Well thats suspicious...</code>
      * 
      * @param logger server log logger
@@ -91,14 +105,26 @@ public class JobContext {
      */
     public void logWarn(Logger logger, String message) {
         logger.warn(message);
+        logWarn(message);
+    }
+
+    /**
+     * Adds a timestamp followed by an warn marker and the warn message as a new line to the executions log<br>
+     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
+     * Error marker: <code>[WARN]</code> or as defined in {@link JobEngineConfig#LOG_WARN_MARKER}<br>
+     * Example: <code>[22:06:42.680] [WARN] Well thats suspicious...</code>
+     * 
+     * @param message text to log
+     */
+    public void logWarn(String message) {
         appendLog(message, true, "w");
     }
 
     /**
      * Adds a timestamp followed by an error marker and the error message as a new line to the executions log. It also adds the message in severity ERROR to the
      * server log<br>
-     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code><br>
-     * Error marker: <code>[ERROR]</code><br>
+     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
+     * Error marker: <code>[ERROR]</code> or as defined in {@link JobEngineConfig#LOG_ERROR_MARKER}<br>
      * Example: <code>[22:06:42.680] [ERROR] Dafuq was that?!?!</code>
      * 
      * @param logger server log logger
@@ -106,6 +132,18 @@ public class JobContext {
      */
     public void logError(Logger logger, String message) {
         logger.error(message);
+        logError(message);
+    }
+
+    /**
+     * Adds a timestamp followed by an error marker and the error message as a new line to the executions log<br>
+     * Timestamp pattern: <code>[HH:mm:ss.SSS]</code> or as defined in {@link JobEngineConfig#LOG_TIME_FORMATTER}<br>
+     * Error marker: <code>[ERROR]</code> or as defined in {@link JobEngineConfig#LOG_ERROR_MARKER}<br>
+     * Example: <code>[22:06:42.680] [ERROR] Dafuq was that?!?!</code>
+     * 
+     * @param message text to log
+     */
+    public void logError(String message) {
         appendLog(message, true, "e");
     }
 
@@ -115,16 +153,27 @@ public class JobContext {
             logBuffer.append(System.lineSeparator());
         }
         if (timestamp) {
-            logBuffer.append("[");
-            logBuffer.append(JobEngineUtil.timestamp().toLocalTime());
-            logBuffer.append("] ");
+            logBuffer.append(JobEngineUtil.timestamp().format(JobEngineConfig.LOG_TIME_FORMATTER));
+            logBuffer.append(" ");
         }
         switch (mode) {
+            case "i":
+                if (JobEngineConfig.LOG_INFO_MARKER != null) {
+                    logBuffer.append(JobEngineConfig.LOG_INFO_MARKER);
+                    logBuffer.append(" ");
+                }
+                break;
             case "w":
-                logBuffer.append("[WARN] ");
+                if (JobEngineConfig.LOG_WARN_MARKER != null) {
+                    logBuffer.append(JobEngineConfig.LOG_WARN_MARKER);
+                    logBuffer.append(" ");
+                }
                 break;
             case "e":
-                logBuffer.append("[ERROR] ");
+                if (JobEngineConfig.LOG_ERROR_MARKER != null) {
+                    logBuffer.append(JobEngineConfig.LOG_ERROR_MARKER);
+                    logBuffer.append(" ");
+                }
                 break;
             default:
                 break;