+++ src/main/java/io/coodoo/workhorse/jobengine/boundary/JobEngineConfig.java
@@ -1,0 +1,57 @@
+package io.coodoo.workhorse.jobengine.boundary;
+
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+
+/**
+ * Basic configuration that can be changed in the implementation<br>
+ * If for example you want to change the error log marker, just call<br>
+ * <tt>JobEngineConfig.LOG_ERROR_MARKER = "SCREW-UP:"</tt>
+ * 
+ * @author coodoo GmbH (coodoo.io)
+ */
+public final class JobEngineConfig {
+
+    private JobEngineConfig() {}
+
+    /**
+     * ZoneId Object time zone for LocalDateTime instance creation. Default is UTC
+     */
+    public static ZoneId TIME_ZONE = ZoneId.of("UTC");
+
+    /**
+     * Job queue poller interval in seconds
+     */
+    public static int JOB_QUEUE_POLLER_INTERVAL = 5;
+
+    /**
+     * Max amount of executions to load into the memory queue per job
+     */
+    public static int JOB_QUEUE_MAX = 1000;
+
+    /**
+     * Min amount of executions in memory queue before the poller gets to add more
+     */
+    public static int JOB_QUEUE_MIN = 100;
+
+    /**
+     * Log timestamp pattern. Default is <code>[HH:mm:ss.SSS]</code>
+     */
+    public static DateTimeFormatter LOG_TIME_FORMATTER = DateTimeFormatter.ofPattern("'['HH:mm:ss.SSS']'");
+
+    /**
+     * Log info marker. Default is none
+     */
+    public static String LOG_INFO_MARKER = null;
+
+    /**
+     * Log warn marker. Default is <code>[WARN]</code>
+     */
+    public static String LOG_WARN_MARKER = "[WARN]";
+
+    /**
+     * Log error marker. Default is <code>[ERROR]</code>
+     */
+    public static String LOG_ERROR_MARKER = "[ERROR]";
+
+}